+++ Physics Engine/Physics Engine/Spring.h
@@ -6,14 +6,14 @@ class Spring
 	{
 	private:
 		sf::Sprite spring_sprite;
-		float scale_const = 1;
+		double scale_const = 1;
 
 		Vectord right, left, Force;
-		float hardness = 1;
-		float init_dist = 0, curr_dist = 0, delta_dist = 0,
+		double hardness = 1;
+		double init_dist = 0, curr_dist = 0, delta_dist = 0,
 			  max_dist  = 0, min_dist  = 0;
 	public:
-		Spring (sf::Sprite* spring_spr, Vectord begin, Vectord end, float k);
+		Spring (sf::Sprite* spring_spr, Vectord begin, Vectord end, double k);
 		~Spring ();
 
 		Vectord getForceRight ();
@@ -22,5 +22,5 @@ class Spring
 		void update (Vectord begin, Vectord end);
 		void draw (sf::RenderWindow &window, bool points = false);
 
-        float getPotEnergy ();
+        double getPotEnergy ();
 	};
